codebaseGenius/requirements.txt
streamlit>=1.39.0
python-dotenv>=1.0.1
google-genai>=0.1.0
